Move public include file.
