+++ libs/tkrng/test_package/example.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "RNG.hpp"
+#include "tkrng/RNG.hpp"
 #include <iostream>
 
 using namespace tket;
